ms-metadata/code/requirements.txt
requests~=2.31.0
python-dotenv~=0.19.1
minio~=7.2.0
pika~=1.3.2
pyspf~=2.0.14
py3dns~=4.0.0
mail-parser~=3.15.0
